Added missing LocalFileClient that was skipped by Git, for some reason.
